move component
this component will be shared, so move it to a more high-level place
